requirements.txt
streamlit
torch
transformers
plotly
groq
python-dotenv
pandas